Optionally allow agent forwarding with sudo
This patch adds a new option: ANSIBLE_ALLOW_AGENT_FORWARDING_WITH_SUDO

When this option is enabled, agent forwarding will work when sudo'ing,
if using the ssh connection. By default, the option is disabled.

The use case that motivated this is to enable using the git module
with sudo to clone a private repository.

If the sudo_user is not root, and the option is enabled, then the
permissions on the ssh auth socket will be modified to allow access
using acls.

This feature required modifying the make_sudo_cmd in order to allow
passing of environment variables (in this case, SSH_AUTH_SOCK), from
user to sudo_user.